+++ src/storage/trunk_freelist.c
@@ -179,7 +179,7 @@ static int waiting_avail_trunk(struct fs_trunk_allocator *allocator,
 }
 
 int trunk_freelist_alloc_space(struct fs_trunk_allocator *allocator,
-        FSTrunkFreelist *freelist, const uint32_t blk_hc, const int size,
+        FSTrunkFreelist *freelist, const uint64_t blk_hc, const int size,
         FSTrunkSpaceWithVersion *spaces, int *count, const bool is_normal)
 {
     int aligned_size;
